fix(tests): fix 4 CI failures from config-clients cycle refactor
The refactor in 7ea04734 moved git helpers from clients to utils but
left a circular dependency (utils→config via codeagent_helpers) and
broke test mock paths pointing to the old location.

- Extract config-dependent prompt helpers to agents/codeagent_prompts.py
- Update all test mock paths from vibe3.clients.git_client to vibe3.utils.git_helpers
- Clear lru_cache on find_repo_root in tests for proper isolation
