+++ src/vibe3/utils/codeagent_helpers.py
@@ -1,29 +1,17 @@
-"""Helper functions for Codeagent backend."""
+"""Helper functions for Codeagent backend.
+
+Pure utility functions — no config dependency (lives in layer 6 utils).
+Config-dependent prompt helpers moved to vibe3.agents.codeagent_prompts.
+"""
 
 from __future__ import annotations
 
 import re
-import tempfile
-from pathlib import Path
-from typing import TYPE_CHECKING, Any, Final
+from typing import Any, Final
 
 from loguru import logger
 
 from vibe3.utils.constants import CODEAGENT_STDIN_MODE_THRESHOLD
-
-# Delayed import to avoid utils → config circular dependency
-# from vibe3.config import VibeConfig
-
-if TYPE_CHECKING:
-    from vibe3.config import VibeConfig
-
-
-def get_vibe_config() -> "VibeConfig":
-    """Get VibeConfig with delayed import."""
-    from vibe3.config import VibeConfig
-
-    return VibeConfig.get_defaults()
-
 
 # Known backend-internal error patterns with suggested fixes
 KNOWN_BACKEND_ERROR_PATTERNS: Final[tuple[tuple[str, str, str], ...]] = (
@@ -151,36 +139,8 @@ def summarize_backend_output(stderr: str, stdout: str) -> str:
     return preview[:500]
 
 
-def build_prompt_file_content(prompt: str, include_global_notice: bool = True) -> str:
-    """Apply configured global notice to the prompt file content."""
-    if not include_global_notice:
-        return prompt
-    config = get_vibe_config()
-    notice = config.agent_prompt.global_notice.strip()
-    if not notice:
-        return prompt
-    return f"{notice}\n\n---\n\n{prompt}"
-
-
-def prepare_prompt_file(
-    prompt: str, include_global_notice: bool = True
-) -> tuple[Path, str]:
-    """Create temporary prompt file with global notice.
-
-    Returns:
-        tuple of (file_path, prompt_content) to avoid caller needing to
-        call build_prompt_file_content separately for diagnostics.
-    """
-    prompt_dir = Path.home() / ".codeagent" / "agents"
-    prompt_dir.mkdir(parents=True, exist_ok=True)
-    prompt_content = build_prompt_file_content(
-        prompt, include_global_notice=include_global_notice
-    )
-    with tempfile.NamedTemporaryFile(
-        mode="w", suffix=".md", delete=False, dir=prompt_dir
-    ) as f:
-        f.write(prompt_content)
-        return Path(f.name), prompt_content
+# build_prompt_file_content and prepare_prompt_file moved to
+# vibe3.agents.codeagent_prompts to break utils → config cycle.
 
 
 def sanitize_task_shell_meta(task: str) -> str:
