+++ tests/vibe3/agents/test_codeagent_backend_resume.py
@@ -7,7 +7,7 @@ from vibe3.models.review_runner import AgentOptions
 
 
 @patch.object(CodeagentBackend, "_run_subprocess")
-@patch("vibe3.utils.codeagent_helpers.tempfile.NamedTemporaryFile")
+@patch("vibe3.agents.codeagent_prompts.tempfile.NamedTemporaryFile")
 def test_codeagent_backend_resume_mode(mock_tempfile, mock_run):
     mock_file = MagicMock()
     mock_file.name = "/Users/test/.codeagent/agents/fake-prompt.md"
@@ -48,7 +48,7 @@ def test_codeagent_backend_resume_mode(mock_tempfile, mock_run):
 
 
 @patch.object(CodeagentBackend, "_run_subprocess")
-@patch("vibe3.utils.codeagent_helpers.tempfile.NamedTemporaryFile")
+@patch("vibe3.agents.codeagent_prompts.tempfile.NamedTemporaryFile")
 def test_codeagent_backend_new_session(mock_tempfile, mock_run):
     mock_file = MagicMock()
     mock_file.name = "/Users/test/.codeagent/agents/fake-prompt.md"
